backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.8.3
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
fastapi==0.104.1
fonttools==4.60.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
jiter==0.11.0
kiwisolver==1.4.9
matplotlib==3.10.6
numpy==2.3.3
openai==1.109.1
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pydantic==2.11.9
pydantic_core==2.33.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.3
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
